Fix issue with ListUnits parser
* Fix #2243 problem with parsing headers as values
* Add tests for data with headers
* Add doctests for both parsers in module unitfiles and incorporate all
  module tests in test_unitfiles module.  Remove test_listunits.
* Refactor logic for ListUnits parser
